Disable browser caching for BPMN using $wgResourceLoaderDebug
There may be a better way to do this, but this works. The combination of bpmn-js and ResourceLoader causes old versions of diagrams, in both view and edit mode, to be cached in the browser unless a "hard refresh" is done, which can lead to a variety of issues. Turning on debug mode gets rid of the caching.

Change-Id: Ifafde614bd942d26bc9ce1c58f5407054568c136